Disable non-TLS accès to linuxfr.org (redirect port 80 to 443)
This would be more secure for everyone as it would prevent accidental
non-HTTPS authentication of users and the associated leaking of the
username and password of the users to potential eavesdroppers.

In this day and age, is is very likely everyone can access HTTPS
websites as many popular websites mandate access through HTTPS.

The "dlfp.lo" configuration file was left untouched because it was
already not up to date.

Suivi #1808 - Administration système Enlever l'accès sans TLS?
(rediriger http://linuxfr.org vers https://linuxfr.org)